feat(recorder): extract recorder as a module
